fix: PRレビューのフィードバックを反映
- FakeLineApiClientをutils.tsから専用ファイルに分離
- supabase_integration_test.yamlにpermissions: contents: readを追加
- e2e_test.yamlのステップ名をE2Eに統一
